docs: Add ResponsiveBrowser to RizzyUI documentation
Added a new component, ResponsiveBrowser, to the RizzyUI documentation. This includes creating new files for the component and its child content property, as well as updating existing files to reflect this addition. Also updated the Home page in the docs to demonstrate usage of the ResponsiveBrowser component. Lastly, added some CSS rules related to this change.