feat: add Divider component and documentation
Added a new UI component, Divider, to the RizzyUI library. This includes various styles such as solid, dashed, and dotted dividers. Also added options for labels with different alignments (start, center, end) and an option for an icon label. Accompanying documentation was created in the RizzyUI.Docs project to demonstrate usage of the Divider component.