feat: update RizzyUI and target frameworks
This commit updates the RizzyUI library to version 4.0.1, simplifies the service addition in `Program.cs`, and adds support for .NET 9.0 in addition to .NET 8.0 in the target frameworks of `RizzyUI.csproj`. It also includes an upgrade of TailwindCSS from v3.4.15 to v3.4.17 in `rizzyui.css` and a minor change in `FieldLabel.razor` where a new namespace is used.