feat: Add MarkdownPipeline to Markdown component
A new static property `MarkdownPipeline` has been added to the `Markdown` component. This pipeline is initialized in the `OnInitialized` method using a `MarkdownPipelineBuilder`. The pipeline is then used when converting markdown text to HTML, replacing the previous direct conversion.

Additionally, updated the version of TailwindMerge.NET from 0.3.0 to 1.0.0 in RizzyUI.csproj file dependencies.