feat(RizzyUI): update Rizzy package and modify script references
This commit updates the Rizzy package from version 5.0.6 to 5.0.7 in both RizzyUI.Docs and RizzyUI projects. It also modifies the script reference in App.razor of the RizzyUI.Docs project, adding a new reference to rizzyui.js file.

In addition, it includes an update to llmquery.prompt file with a new generated timestamp and added content related to wwwroot/js/rizzyui.js.

Lastly, the version of RizzyUI is updated from 0.6.1-alpha to 0.6.2-alpha.