docs: revamp README for RizzyUI library
This commit revamps the README for the RizzyUI ASP.NET component library. The changes include the introduction of emoji, modification of section headers and content for readability, the addition of a library logo, and restructured paragraphs. This reorganization aims to make the README more approachable and engaging, thus providing clearer insights into the library features and benefits.